Split document from base.
