+++ sed/transpiler/importers/base.py
@@ -1,7 +1,5 @@
 from pathlib import Path
 from typing import Any
-from sed.transpiler.importers.tasks import load_tasks_section
-from sed.transpiler.importers.outputs import load_outputs_section
 
 import pandas as pd
 
@@ -50,37 +48,3 @@ class Range(SedBase):
             print("Unsaved data when creating Range:", leftovers)
             return True
         return False
-
-class SedDocument():
-    """The document itself."""
-    def __init__(self, config: dict):
-        self.versionStr = config.pop("versionStr", None)
-        self.versionNum = config.pop("versionNum", None)
-        self.constants = config.pop("constants", None)
-        self.tasks = load_tasks_section(config.pop("tasks", None))
-        self.outputs = load_outputs_section(config.pop("outputs", None))
-        self.validate(config)
-
-    def validate(self, leftovers={}):
-        """Validate."""
-        if len(leftovers):
-            print("Unsaved data when creating SEDDocument:", leftovers)
-            return True
-        return False
-
-    def exportToPython(self, context, path):
-        headers = set()
-        python = "# Translation of SED document v" + sed.versionStr + " to python\n\n"
-        python +=  "\n# tasks:" + task_key + "\n"
-        for task in self.tasks:
-            newheaders, newpython = task.exportToPython()
-            headers.update(newheaders)
-            python += newpython
-        python +=  "\n# outputs:" + task_key + "\n"
-        for output in self.outputs:
-            newheaders, newpython = output.exportToPython()
-            headers.update(newheaders)
-            python += newpython
-        return headers, python
-
-
